Committed script, readme, and zip of disk images
This upload has everything one needs to start using the hard disk support on the Ampro Z80 Little Board (Plus) simulator in MAME.  It needs fully-preallocated disk images filled with the magic value E5 to indicate that wherever a CP/M directory falls, it'll see an empty directory.  The upload consists of the script I used to create the images, the README that's in the zip file describing the files within, as well as the devices' geometry.  The script also has documentation in the comments on how to create the resulting CHD file once a raw E5-filled file is created.